+++ AdProject.Domain/Entidades/SubCategoria.cs
@@ -4,9 +4,9 @@ using System.Text;
 
 namespace AdProject.Dominio.Entidades
 {
-    public class Subcategoria : Entidade<long>
+    public class Subcategoria : Entidade<int>
     {
-        public long IdCategoria { get; set; }
+        public int IdCategoria { get; set; }
         public string Nome { get; set; }
         public virtual Categoria Categoria { get; set; }
     }
